Conflicts:
	projects/com.oracle.truffle.llvm.test/interoptests/interop028.c
	projects/com.oracle.truffle.llvm.test/interoptests/interop029.c
	projects/com.oracle.truffle.llvm.test/src/com/oracle/truffle/llvm/test/interop/LLVMInteropTest.java

+++ projects/com.oracle.truffle.llvm.test/interoptests/interop033.c
@@ -1,2 +1,2 @@
-long get(long *a, int idx) { return a[idx]; }
+int get(short *a, int idx) { return a[idx]; }
 int main() { return 0; }
