+++ projects/com.oracle.truffle.llvm.test/interoptests/interop035.c
@@ -1,2 +1,2 @@
-double get(double *a, int idx) { return a[idx]; }
+long get(long *a, int idx) { return a[idx]; }
 int main() { return 0; }
