+++ projects/com.oracle.truffle.llvm.test/interoptests/interop036.c
@@ -1,2 +1,2 @@
-int get(int *a, int idx) { return a[idx]; }
+float get(float *a, int idx) { return a[idx]; }
 int main() { return 0; }
